Move calculate_aligned_bases to utils
