+++ src/utils.cpp
@@ -1,3 +1,7 @@
+#include <htslib/sam.h>
+#include <string>
+#include <vector>
+
 //' Get unmodified base corresponding to a modified base
 //'
 //' @param b Modified base as a char
@@ -65,3 +69,20 @@ char complement(char n) {
     }
 }
 
+// calculate aligned bases (sum of 'M', '=', or 'X' operation lengths)
+int calculate_aligned_bases(bam1_t *bamdata) {
+    uint32_t *cigar = bam_get_cigar(bamdata);
+    int aligned_bases = 0;
+    
+    // loop over CIGAR operations
+    for (uint32_t i = 0; i < bamdata->core.n_cigar; i++) {
+        uint32_t op = bam_cigar_op(cigar[i]);
+        
+        // only count 'M', '=', or 'X' operations
+        if (op == BAM_CMATCH || op == BAM_CEQUAL || op == BAM_CDIFF) {
+            aligned_bases += bam_cigar_oplen(cigar[i]);
+        }
+    }
+    
+    return aligned_bases;
+}
