+++ src/utils.h
@@ -1,2 +1,5 @@
+#include <htslib/sam.h>
+
 char get_unmodified_base(char);
 char complement(char);
+int calculate_aligned_bases(bam1_t*);
